fix(lora): race-free LoRAManager singleton via OnceLock
Centralizes LoRAInfo and get_lora_manager() in
common.lora.manager and switches the vllm and sglang
handlers to import from there, removing their local
copies. The new get_lora_manager() uses OnceLock for
race-free lazy initialization, so concurrent calls no
longer construct multiple LoRAManager instances with
independent download caches.

Refs: DIS-1840
